feat: add "recovered" column to table output
this checks the list of recently recovered services against the listed
notifications to see if there was a recovery after the notification has
been sent.

This might not work correctly in some edge cases (lots of recoveries
for services without any notifications so that the recoveries for the
last few notifications are already rotated out) but should usually be
helpful for most use cases.

I couldn't find a better way to solve this task without sending a ton of
requests to icinga...